requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
cachetools==5.3.3
    # via google-auth
certifi==2024.2.2
    # via requests
charset-normalizer==3.3.2
    # via requests
google-api-core==2.19.0
    # via google-api-python-client
google-api-python-client==2.131.0
    # via -r requirements.in
google-auth==2.29.0
    # via
    #   -r requirements.in
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-auth-oauthlib
    #   gspread
google-auth-httplib2==0.2.0
    # via
    #   -r requirements.in
    #   google-api-python-client
google-auth-oauthlib==1.2.0
    # via
    #   -r requirements.in
    #   gspread
googleapis-common-protos==1.63.0
    # via google-api-core
gspread==6.1.2
    # via -r requirements.in
httplib2==0.22.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
idna==3.7
    # via requests
oauthlib==3.2.2
    # via requests-oauthlib
proto-plus==1.23.0
    # via google-api-core
protobuf==4.25.3
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   proto-plus
pyasn1==0.6.0
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.0
    # via google-auth
pyparsing==3.1.2
    # via httplib2
python-dateutil==2.9.0.post0
    # via -r requirements.in
python-dotenv==1.0.1
    # via -r requirements.in
requests==2.32.3
    # via
    #   -r requirements.in
    #   google-api-core
    #   requests-oauthlib
requests-oauthlib==2.0.0
    # via google-auth-oauthlib
rsa==4.9
    # via google-auth
six==1.16.0
    # via python-dateutil
slack-sdk==3.27.2
    # via -r requirements.in
uritemplate==4.1.1
    # via google-api-python-client
urllib3==2.2.1
    # via requests
